Simplify force implementation.
For consistency, we always compute an acceleration vector on the compute
device, and only transform it into a force vector in the code that
modifies the mass fractions (by having density in the force prefactor).

This commit also fixes a bug which caused the force implementation to
use a spurious 'rho' factor in non-Shan-Chen simulations.